+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 
 v5.0.0
 ----
-* Support Capacitor Projects
+* Support Capacitor Projects (Beta).
 
 v4.2.3
 ----